Create motion activated lighting
This automation is used to turn lights in a specific room on when there is motion. It also will turn the lights off in the room after a period of no motion.  There is a condition that was added to keep the lights from turning on when I am asleep.